Изменение структуры проекта
Along the way, it was decided to radically redesign the site's system and redesign the design of the site, making it more pleasing to the eye and adding a couple of new features that will make it individual. To do this, I decided to move the old project to another branch so that it would not interfere with the main site, which would run on react rather than regular html and css. Due to the fact that I initially started writing in plain html and tried to combine technologies, this caused delays in development.